Separated most of the parsing routines into files
